Big Y Updated
Updated the tool to reflect new salary.

Changed the display of the graph. Originally it showed the % as percentage of specific cost to all costs. Changed this to show specific cost in relation to amount earned.

Changed the difference to reflect surplus/deficit on a specific cost to ideal specific cost